+++ .gitignore
@@ -83,7 +83,7 @@ profile_default/
 ipython_config.py
 
 # pyenv
-#   For a library or package, you might want to ignore these files since the code is
+#   For a library or package, you might want to ignore these files since the code_account is
 #   intended to run in multiple environments; otherwise, check them in:
 # .python-version
