+++ code_account/main.py
@@ -1,0 +1,3 @@
+from code_account.operations_func import print_of_operations, sort_operations, unpacking_json
+
+print(print_of_operations(sort_operations(unpacking_json('operations.json'))))
